With error boundry (and unbundled Preact)
