+++ frontend/components/CellInput.js
@@ -351,10 +351,14 @@ export const CellInput = ({
     set_show_logs,
     cm_highlighted_line,
     variables_in_all_notebook,
-    onerror,
 }) => {
     let pluto_actions = useContext(PlutoContext)
-
+    let [error, set_error] = useState(null)
+    if (error) {
+        const to_throw = error
+        set_error(null)
+        throw to_throw
+    }
     const newcm_ref = useRef(/** @type {EditorView} */ (null))
     const dom_node_ref = useRef(/** @type {HTMLElement} */ (null))
     const remote_code_ref = useRef(null)
@@ -661,8 +665,7 @@ export const CellInput = ({
                     // debug_syntax_plugin,
                     // Handle errors hopefully?
                     EditorView.exceptionSink.of((exception) => {
-                        console.error(`Exception Sink captured an exception and will reload codemirror:\n${exception}`)
-                        onerror()
+                        set_error(exception)
                         alert(
                             "We run into an issue; We have lost your cursor 😞😓😿\n If this appears again, please report an issue at https://github.com/fonsp/Pluto.jl/issues"
                         )
