+++ frontend/imports/Preact.js
@@ -1,544 +1,47 @@
-var te = Object.defineProperty;
-var ne = (e, t)=>{
-    for(var n in t)te(e, n, {
-        get: t[n],
-        enumerable: !0
-    });
-};
-var N = {
-};
-ne(N, {
-    Component: ()=>S
-    ,
-    Fragment: ()=>P
-    ,
-    cloneElement: ()=>se
-    ,
-    createContext: ()=>ce
-    ,
-    createElement: ()=>V
-    ,
-    createRef: ()=>re
-    ,
-    h: ()=>V
-    ,
-    hydrate: ()=>ee
-    ,
-    isValidElement: ()=>F
-    ,
-    options: ()=>d
-    ,
-    render: ()=>Z
-    ,
-    toChildArray: ()=>K
-});
-var d, F, C, H, R, I, U = {
-}, L = [], _e = /acit|ex(?:s|g|n|p|$)|rph|grid|ows|mnc|ntw|ine[ch]|zoo|^ord|itera/i;
-function b(e, t) {
-    for(var n in t)e[n] = t[n];
-    return e;
+// @ts-nocheck
+import {
+    //
+    render,
+    Component,
+    h,
+    cloneElement,
+    createContext,
+    createRef,
+    hydrate,
+} from "https://esm.sh/preact@10.6.6?target=es2020"
+
+import {
+    //
+    useEffect,
+    useLayoutEffect,
+    useState,
+    useRef,
+    useMemo,
+    useCallback,
+    useContext,
+    useErrorBoundary,
+} from "https://esm.sh/preact@10.6.6/hooks?target=es2020"
+
+import htm from "https://esm.sh/htm@3.0.4?target=es2020"
+
+const html = htm.bind(h)
+
+export {
+    //
+    html,
+    render,
+    Component,
+    useEffect,
+    useLayoutEffect,
+    useState,
+    useRef,
+    useMemo,
+    useCallback,
+    useErrorBoundary,
+    createContext,
+    createRef,
+    useContext,
+    h,
+    cloneElement,
+    hydrate,
 }
-function O(e) {
-    var t = e.parentNode;
-    t && t.removeChild(e);
-}
-function V(e, t, n) {
-    var o, l, _, c = arguments, s = {
-    };
-    for(_ in t)_ == "key" ? o = t[_] : _ == "ref" ? l = t[_] : s[_] = t[_];
-    if (arguments.length > 3) for(n = [
-        n
-    ], _ = 3; _ < arguments.length; _++)n.push(c[_]);
-    if (n != null && (s.children = n), typeof e == "function" && e.defaultProps != null) for(_ in e.defaultProps)s[_] === void 0 && (s[_] = e.defaultProps[_]);
-    return x(e, s, o, l, null);
-}
-function x(e, t, n, o, l) {
-    var _ = {
-        type: e,
-        props: t,
-        key: n,
-        ref: o,
-        __k: null,
-        __: null,
-        __b: 0,
-        __e: null,
-        __d: void 0,
-        __c: null,
-        __h: null,
-        constructor: void 0,
-        __v: l ?? ++d.__v
-    };
-    return d.vnode != null && d.vnode(_), _;
-}
-function re() {
-    return {
-        current: null
-    };
-}
-function P(e) {
-    return e.children;
-}
-function S(e, t) {
-    this.props = e, this.context = t;
-}
-function E(e, t) {
-    if (t == null) return e.__ ? E(e.__, e.__.__k.indexOf(e) + 1) : null;
-    for(var n; t < e.__k.length; t++)if ((n = e.__k[t]) != null && n.__e != null) return n.__e;
-    return typeof e.type == "function" ? E(e) : null;
-}
-function $(e) {
-    var t, n;
-    if ((e = e.__) != null && e.__c != null) {
-        for(e.__e = e.__c.base = null, t = 0; t < e.__k.length; t++)if ((n = e.__k[t]) != null && n.__e != null) {
-            e.__e = e.__c.base = n.__e;
-            break;
-        }
-        return $(e);
-    }
-}
-function M(e) {
-    (!e.__d && (e.__d = !0) && C.push(e) && !T.__r++ || R !== d.debounceRendering) && ((R = d.debounceRendering) || H)(T);
-}
-function T() {
-    for(var e; T.__r = C.length;)e = C.sort(function(t, n) {
-        return t.__v.__b - n.__v.__b;
-    }), C = [], e.some(function(t) {
-        var n, o, l, _, c, s;
-        t.__d && (c = (_ = (n = t).__v).__e, (s = n.__P) && (o = [], (l = b({
-        }, _)).__v = _.__v + 1, W(s, _, l, n.__n, s.ownerSVGElement !== void 0, _.__h != null ? [
-            c
-        ] : null, o, c ?? E(_), _.__h), B(o, _), _.__e != c && $(_)));
-    });
-}
-function J(e, t, n, o, l, _, c, s, p, u) {
-    var r, v, f, i, y, a, h, m = o && o.__k || L, g = m.length;
-    for(n.__k = [], r = 0; r < t.length; r++)if ((i = n.__k[r] = (i = t[r]) == null || typeof i == "boolean" ? null : typeof i == "string" || typeof i == "number" || typeof i == "bigint" ? x(null, i, null, null, i) : Array.isArray(i) ? x(P, {
-        children: i
-    }, null, null, null) : i.__b > 0 ? x(i.type, i.props, i.key, null, i.__v) : i) != null) {
-        if (i.__ = n, i.__b = n.__b + 1, (f = m[r]) === null || f && i.key == f.key && i.type === f.type) m[r] = void 0;
-        else for(v = 0; v < g; v++){
-            if ((f = m[v]) && i.key == f.key && i.type === f.type) {
-                m[v] = void 0;
-                break;
-            }
-            f = null;
-        }
-        W(e, i, f = f || U, l, _, c, s, p, u), y = i.__e, (v = i.ref) && f.ref != v && (h || (h = []), f.ref && h.push(f.ref, null, i), h.push(v, i.__c || y, i)), y != null ? (a == null && (a = y), typeof i.type == "function" && i.__k != null && i.__k === f.__k ? i.__d = p = j(i, p, e) : p = z(e, i, f, m, y, p), u || n.type !== "option" ? typeof n.type == "function" && (n.__d = p) : e.value = "") : p && f.__e == p && p.parentNode != e && (p = E(f));
-    }
-    for(n.__e = a, r = g; r--;)m[r] != null && (typeof n.type == "function" && m[r].__e != null && m[r].__e == n.__d && (n.__d = E(o, r + 1)), q(m[r], m[r]));
-    if (h) for(r = 0; r < h.length; r++)G(h[r], h[++r], h[++r]);
-}
-function j(e, t, n) {
-    var o, l;
-    for(o = 0; o < e.__k.length; o++)(l = e.__k[o]) && (l.__ = e, t = typeof l.type == "function" ? j(l, t, n) : z(n, l, l, e.__k, l.__e, t));
-    return t;
-}
-function K(e, t) {
-    return t = t || [], e == null || typeof e == "boolean" || (Array.isArray(e) ? e.some(function(n) {
-        K(n, t);
-    }) : t.push(e)), t;
-}
-function z(e, t, n, o, l, _) {
-    var c, s, p;
-    if (t.__d !== void 0) c = t.__d, t.__d = void 0;
-    else if (n == null || l != _ || l.parentNode == null) e: if (_ == null || _.parentNode !== e) e.appendChild(l), c = null;
-    else {
-        for(s = _, p = 0; (s = s.nextSibling) && p < o.length; p += 2)if (s == l) break e;
-        e.insertBefore(l, _), c = _;
-    }
-    return c !== void 0 ? c : l.nextSibling;
-}
-function oe(e, t, n, o, l) {
-    var _;
-    for(_ in n)_ === "children" || _ === "key" || _ in t || D(e, _, null, n[_], o);
-    for(_ in t)l && typeof t[_] != "function" || _ === "children" || _ === "key" || _ === "value" || _ === "checked" || n[_] === t[_] || D(e, _, t[_], n[_], o);
-}
-function Q(e, t, n) {
-    t[0] === "-" ? e.setProperty(t, n) : e[t] = n == null ? "" : typeof n != "number" || _e.test(t) ? n : n + "px";
-}
-function D(e, t, n, o, l) {
-    var _;
-    e: if (t === "style") if (typeof n == "string") e.style.cssText = n;
-    else {
-        if (typeof o == "string" && (e.style.cssText = o = ""), o) for(t in o)n && t in n || Q(e.style, t, "");
-        if (n) for(t in n)o && n[t] === o[t] || Q(e.style, t, n[t]);
-    }
-    else if (t[0] === "o" && t[1] === "n") _ = t !== (t = t.replace(/Capture$/, "")), t = t.toLowerCase() in e ? t.toLowerCase().slice(2) : t.slice(2), e.l || (e.l = {
-    }), e.l[t + _] = n, n ? o || e.addEventListener(t, _ ? Y : X, _) : e.removeEventListener(t, _ ? Y : X, _);
-    else if (t !== "dangerouslySetInnerHTML") {
-        if (l) t = t.replace(/xlink[H:h]/, "h").replace(/sName$/, "s");
-        else if (t !== "href" && t !== "list" && t !== "form" && t !== "tabIndex" && t !== "download" && t in e) try {
-            e[t] = n ?? "";
-            break e;
-        } catch (c) {
-        }
-        typeof n == "function" || (n != null && (n !== !1 || t[0] === "a" && t[1] === "r") ? e.setAttribute(t, n) : e.removeAttribute(t));
-    }
-}
-function X(e) {
-    this.l[e.type + !1](d.event ? d.event(e) : e);
-}
-function Y(e) {
-    this.l[e.type + !0](d.event ? d.event(e) : e);
-}
-function W(e, t, n, o, l, _, c, s, p) {
-    var u, r, v, f, i, y, a, h, m, g, w, k = t.type;
-    if (t.constructor !== void 0) return null;
-    n.__h != null && (p = n.__h, s = t.__e = n.__e, t.__h = null, _ = [
-        s
-    ]), (u = d.__b) && u(t);
-    try {
-        e: if (typeof k == "function") {
-            if (h = t.props, m = (u = k.contextType) && o[u.__c], g = u ? m ? m.props.value : u.__ : o, n.__c ? a = (r = t.__c = n.__c).__ = r.__E : ("prototype" in k && k.prototype.render ? t.__c = r = new k(h, g) : (t.__c = r = new S(h, g), r.constructor = k, r.render = ie), m && m.sub(r), r.props = h, r.state || (r.state = {
-            }), r.context = g, r.__n = o, v = r.__d = !0, r.__h = []), r.__s == null && (r.__s = r.state), k.getDerivedStateFromProps != null && (r.__s == r.state && (r.__s = b({
-            }, r.__s)), b(r.__s, k.getDerivedStateFromProps(h, r.__s))), f = r.props, i = r.state, v) k.getDerivedStateFromProps == null && r.componentWillMount != null && r.componentWillMount(), r.componentDidMount != null && r.__h.push(r.componentDidMount);
-            else {
-                if (k.getDerivedStateFromProps == null && h !== f && r.componentWillReceiveProps != null && r.componentWillReceiveProps(h, g), !r.__e && r.shouldComponentUpdate != null && r.shouldComponentUpdate(h, r.__s, g) === !1 || t.__v === n.__v) {
-                    r.props = h, r.state = r.__s, t.__v !== n.__v && (r.__d = !1), r.__v = t, t.__e = n.__e, t.__k = n.__k, t.__k.forEach(function(A) {
-                        A && (A.__ = t);
-                    }), r.__h.length && c.push(r);
-                    break e;
-                }
-                r.componentWillUpdate != null && r.componentWillUpdate(h, r.__s, g), r.componentDidUpdate != null && r.__h.push(function() {
-                    r.componentDidUpdate(f, i, y);
-                });
-            }
-            r.context = g, r.props = h, r.state = r.__s, (u = d.__r) && u(t), r.__d = !1, r.__v = t, r.__P = e, u = r.render(r.props, r.state, r.context), r.state = r.__s, r.getChildContext != null && (o = b(b({
-            }, o), r.getChildContext())), v || r.getSnapshotBeforeUpdate == null || (y = r.getSnapshotBeforeUpdate(f, i)), w = u != null && u.type === P && u.key == null ? u.props.children : u, J(e, Array.isArray(w) ? w : [
-                w
-            ], t, n, o, l, _, c, s, p), r.base = t.__e, t.__h = null, r.__h.length && c.push(r), a && (r.__E = r.__ = null), r.__e = !1;
-        } else _ == null && t.__v === n.__v ? (t.__k = n.__k, t.__e = n.__e) : t.__e = le(n.__e, t, n, o, l, _, c, p);
-        (u = d.diffed) && u(t);
-    } catch (A) {
-        t.__v = null, (p || _ != null) && (t.__e = s, t.__h = !!p, _[_.indexOf(s)] = null), d.__e(A, t, n);
-    }
-}
-function B(e, t) {
-    d.__c && d.__c(t, e), e.some(function(n) {
-        try {
-            e = n.__h, n.__h = [], e.some(function(o) {
-                o.call(n);
-            });
-        } catch (o) {
-            d.__e(o, n.__v);
-        }
-    });
-}
-function le(e, t, n, o, l, _, c, s) {
-    var p, u, r, v, f = n.props, i = t.props, y = t.type, a = 0;
-    if (y === "svg" && (l = !0), _ != null) {
-        for(; a < _.length; a++)if ((p = _[a]) && (p === e || (y ? p.localName == y : p.nodeType == 3))) {
-            e = p, _[a] = null;
-            break;
-        }
-    }
-    if (e == null) {
-        if (y === null) return document.createTextNode(i);
-        e = l ? document.createElementNS("http://www.w3.org/2000/svg", y) : document.createElement(y, i.is && i), _ = null, s = !1;
-    }
-    if (y === null) f === i || s && e.data === i || (e.data = i);
-    else {
-        if (_ = _ && L.slice.call(e.childNodes), u = (f = n.props || U).dangerouslySetInnerHTML, r = i.dangerouslySetInnerHTML, !s) {
-            if (_ != null) for(f = {
-            }, v = 0; v < e.attributes.length; v++)f[e.attributes[v].name] = e.attributes[v].value;
-            (r || u) && (r && (u && r.__html == u.__html || r.__html === e.innerHTML) || (e.innerHTML = r && r.__html || ""));
-        }
-        if (oe(e, i, f, l, s), r) t.__k = [];
-        else if (a = t.props.children, J(e, Array.isArray(a) ? a : [
-            a
-        ], t, n, o, l && y !== "foreignObject", _, c, e.firstChild, s), _ != null) for(a = _.length; a--;)_[a] != null && O(_[a]);
-        s || ("value" in i && (a = i.value) !== void 0 && (a !== e.value || y === "progress" && !a) && D(e, "value", a, f.value, !1), "checked" in i && (a = i.checked) !== void 0 && a !== e.checked && D(e, "checked", a, f.checked, !1));
-    }
-    return e;
-}
-function G(e, t, n) {
-    try {
-        typeof e == "function" ? e(t) : e.current = t;
-    } catch (o) {
-        d.__e(o, n);
-    }
-}
-function q(e, t, n) {
-    var o, l, _;
-    if (d.unmount && d.unmount(e), (o = e.ref) && (o.current && o.current !== e.__e || G(o, null, t)), n || typeof e.type == "function" || (n = (l = e.__e) != null), e.__e = e.__d = void 0, (o = e.__c) != null) {
-        if (o.componentWillUnmount) try {
-            o.componentWillUnmount();
-        } catch (c) {
-            d.__e(c, t);
-        }
-        o.base = o.__P = null;
-    }
-    if (o = e.__k) for(_ = 0; _ < o.length; _++)o[_] && q(o[_], t, n);
-    l != null && O(l);
-}
-function ie(e, t, n) {
-    return this.constructor(e, n);
-}
-function Z(e, t, n) {
-    var o, l, _;
-    d.__ && d.__(e, t), l = (o = typeof n == "function") ? null : n && n.__k || t.__k, _ = [], W(t, e = (!o && n || t).__k = V(P, null, [
-        e
-    ]), l || U, U, t.ownerSVGElement !== void 0, !o && n ? [
-        n
-    ] : l ? null : t.firstChild ? L.slice.call(t.childNodes) : null, _, !o && n ? n : l ? l.__e : t.firstChild, o), B(_, e);
-}
-function ee(e, t) {
-    Z(e, t, ee);
-}
-function se(e, t, n) {
-    var o, l, _, c = arguments, s = b({
-    }, e.props);
-    for(_ in t)_ == "key" ? o = t[_] : _ == "ref" ? l = t[_] : s[_] = t[_];
-    if (arguments.length > 3) for(n = [
-        n
-    ], _ = 3; _ < arguments.length; _++)n.push(c[_]);
-    return n != null && (s.children = n), x(e.type, s, o || e.key, l || e.ref, null);
-}
-function ce(e, t) {
-    var n = {
-        __c: t = "__cC" + I++,
-        __: e,
-        Consumer: function(o, l) {
-            return o.children(l);
-        },
-        Provider: function(o) {
-            var l, _;
-            return this.getChildContext || (l = [], (_ = {
-            })[t] = this, this.getChildContext = function() {
-                return _;
-            }, this.shouldComponentUpdate = function(c) {
-                this.props.value !== c.value && l.some(M);
-            }, this.sub = function(c) {
-                l.push(c);
-                var s = c.componentWillUnmount;
-                c.componentWillUnmount = function() {
-                    l.splice(l.indexOf(c), 1), s && s.call(c);
-                };
-            }), o.children;
-        }
-    };
-    return n.Provider.__ = n.Consumer.contextType = n;
-}
-d = {
-    __e: function(e, t) {
-        for(var n, o, l; t = t.__;)if ((n = t.__c) && !n.__) try {
-            if ((o = n.constructor) && o.getDerivedStateFromError != null && (n.setState(o.getDerivedStateFromError(e)), l = n.__d), n.componentDidCatch != null && (n.componentDidCatch(e), l = n.__d), l) return n.__E = n;
-        } catch (_) {
-            e = _;
-        }
-        throw e;
-    },
-    __v: 0
-}, F = function(e) {
-    return e != null && e.constructor === void 0;
-}, S.prototype.setState = function(e, t) {
-    var n;
-    n = this.__s != null && this.__s !== this.state ? this.__s : this.__s = b({
-    }, this.state), typeof e == "function" && (e = e(b({
-    }, n), this.props)), e && b(n, e), e != null && this.__v && (t && this.__h.push(t), M(this));
-}, S.prototype.forceUpdate = function(e) {
-    this.__v && (this.__e = !0, e && this.__h.push(e), M(this));
-}, S.prototype.render = P, C = [], H = typeof Promise == "function" ? Promise.prototype.then.bind(Promise.resolve()) : setTimeout, T.__r = 0, I = 0;
-var { createElement: ae , cloneElement: ue , toChildArray: de , hydrate: he , Fragment: ve , h: ye , createContext: me , options: ge , createRef: ke , isValidElement: be , Component: Ce , render: xe  } = N;
-var q1 = Object.defineProperty;
-var k = (_, e)=>{
-    for(var n in e)q1(_, n, {
-        get: e[n],
-        enumerable: !0
-    });
-};
-var h1 = {
-};
-k(h1, {
-    useCallback: ()=>S1
-    ,
-    useContext: ()=>B1
-    ,
-    useDebugValue: ()=>I1
-    ,
-    useEffect: ()=>R1
-    ,
-    useErrorBoundary: ()=>L1
-    ,
-    useImperativeHandle: ()=>V1
-    ,
-    useLayoutEffect: ()=>D1
-    ,
-    useMemo: ()=>l
-    ,
-    useReducer: ()=>x1
-    ,
-    useRef: ()=>T1
-    ,
-    useState: ()=>C1
-});
-var c, r, H1, a = 0, m = [], d1 = ge.__b, E1 = ge.__r, y = ge.diffed, b1 = ge.__c, g = ge.unmount;
-function i(_, e) {
-    ge.__h && ge.__h(r, _, a || e), a = 0;
-    var n = r.__H || (r.__H = {
-        __: [],
-        __h: []
-    });
-    return _ >= n.__.length && n.__.push({
-    }), n.__[_];
-}
-function C1(_) {
-    return a = 1, x1(A, _);
-}
-function x1(_, e, n) {
-    var t = i(c++, 2);
-    return t.t = _, t.__c || (t.__ = [
-        n ? n(e) : A(void 0, e),
-        function(o) {
-            var f = t.t(t.__[0], o);
-            t.__[0] !== f && (t.__ = [
-                f,
-                t.__[1]
-            ], t.__c.setState({
-            }));
-        }
-    ], t.__c = r), t.__;
-}
-function R1(_, e) {
-    var n = i(c++, 3);
-    !ge.__s && v(n.__H, e) && (n.__ = _, n.__H = e, r.__H.__h.push(n));
-}
-function D1(_, e) {
-    var n = i(c++, 4);
-    !ge.__s && v(n.__H, e) && (n.__ = _, n.__H = e, r.__h.push(n));
-}
-function T1(_) {
-    return a = 5, l(function() {
-        return {
-            current: _
-        };
-    }, []);
-}
-function V1(_, e, n) {
-    a = 6, D1(function() {
-        typeof _ == "function" ? _(e()) : _ && (_.current = e());
-    }, n == null ? n : n.concat(_));
-}
-function l(_, e) {
-    var n = i(c++, 7);
-    return v(n.__H, e) && (n.__ = _(), n.__H = e, n.__h = _), n.__;
-}
-function S1(_, e) {
-    return a = 8, l(function() {
-        return _;
-    }, e);
-}
-function B1(_) {
-    var e = r.context[_.__c], n = i(c++, 9);
-    return n.__c = _, e ? (n.__ == null && (n.__ = !0, e.sub(r)), e.props.value) : _.__;
-}
-function I1(_, e) {
-    ge.useDebugValue && ge.useDebugValue(e ? e(_) : _);
-}
-function L1(_) {
-    var e = i(c++, 10), n = C1();
-    return e.__ = _, r.componentDidCatch || (r.componentDidCatch = function(t) {
-        e.__ && e.__(t), n[1](t);
-    }), [
-        n[0],
-        function() {
-            n[1](void 0);
-        }
-    ];
-}
-function M1() {
-    m.forEach(function(_) {
-        if (_.__P) try {
-            _.__H.__h.forEach(s), _.__H.__h.forEach(p), _.__H.__h = [];
-        } catch (e) {
-            _.__H.__h = [], ge.__e(e, _.__v);
-        }
-    }), m = [];
-}
-ge.__b = function(_) {
-    r = null, d1 && d1(_);
-}, ge.__r = function(_) {
-    E1 && E1(_), c = 0;
-    var e = (r = _.__c).__H;
-    e && (e.__h.forEach(s), e.__h.forEach(p), e.__h = []);
-}, ge.diffed = function(_) {
-    y && y(_);
-    var e = _.__c;
-    e && e.__H && e.__H.__h.length && (m.push(e) !== 1 && H1 === ge.requestAnimationFrame || ((H1 = ge.requestAnimationFrame) || function(n) {
-        var t, o = function() {
-            clearTimeout(f), F1 && cancelAnimationFrame(t), setTimeout(n);
-        }, f = setTimeout(o, 100);
-        F1 && (t = requestAnimationFrame(o));
-    })(M1)), r = void 0;
-}, ge.__c = function(_, e) {
-    e.some(function(n) {
-        try {
-            n.__h.forEach(s), n.__h = n.__h.filter(function(t) {
-                return !t.__ || p(t);
-            });
-        } catch (t) {
-            e.some(function(o) {
-                o.__h && (o.__h = []);
-            }), e = [], ge.__e(t, n.__v);
-        }
-    }), b1 && b1(_, e);
-}, ge.unmount = function(_) {
-    g && g(_);
-    var e = _.__c;
-    if (e && e.__H) try {
-        e.__H.__.forEach(s);
-    } catch (n) {
-        ge.__e(n, e.__v);
-    }
-};
-var F1 = typeof requestAnimationFrame == "function";
-function s(_) {
-    var e = r;
-    typeof _.__c == "function" && _.__c(), r = e;
-}
-function p(_) {
-    var e = r;
-    _.__c = _.__(), r = e;
-}
-function v(_, e) {
-    return !_ || _.length !== e.length || e.some(function(n, t) {
-        return n !== _[t];
-    });
-}
-function A(_, e) {
-    return typeof e == "function" ? e(_) : e;
-}
-var { useRef: z1 , useMemo: G1 , useDebugValue: J1 , useErrorBoundary: K1 , useReducer: N1 , useLayoutEffect: O1 , useImperativeHandle: Q1 , useEffect: U1 , useCallback: W1 , useContext: X1 , useState: Y1  } = h1;
-var a1 = function(p, o, l, n) {
-    var v;
-    o[0] = 0;
-    for(var u = 1; u < o.length; u++){
-        var f = o[u++], t = o[u] ? (o[0] |= f ? 1 : 2, l[o[u++]]) : o[++u];
-        f === 3 ? n[0] = t : f === 4 ? n[1] = Object.assign(n[1] || {
-        }, t) : f === 5 ? (n[1] = n[1] || {
-        })[o[++u]] = t : f === 6 ? n[1][o[++u]] += t + "" : f ? (v = p.apply(t, a1(p, t, l, [
-            "",
-            null
-        ])), n.push(v), t[0] ? o[0] |= 2 : (o[u - 2] = 0, o[u] = v)) : n.push(t);
-    }
-    return n;
-}, d2 = new Map;
-function w(p) {
-    var o = d2.get(this);
-    return o || (o = new Map, d2.set(this, o)), (o = a1(this, o.get(p) || (o.set(p, o = function(l) {
-        for(var n, v, u = 1, f = "", t = "", g = [
-            0
-        ], c = function(i) {
-            u === 1 && (i || (f = f.replace(/^\s*\n\s*|\s*\n\s*$/g, ""))) ? g.push(0, i, f) : u === 3 && (i || f) ? (g.push(3, i, f), u = 2) : u === 2 && f === "..." && i ? g.push(4, i, 0) : u === 2 && f && !i ? g.push(5, 0, !0, f) : u >= 5 && ((f || !i && u === 5) && (g.push(u, 0, f, v), u = 6), i && (g.push(u, i, 0, v), u = 6)), f = "";
-        }, s = 0; s < l.length; s++){
-            s && (u === 1 && c(), c(s));
-            for(var m = 0; m < l[s].length; m++)n = l[s][m], u === 1 ? n === "<" ? (c(), g = [
-                g
-            ], u = 3) : f += n : u === 4 ? f === "--" && n === ">" ? (u = 1, f = "") : f = n + f[0] : t ? n === t ? t = "" : f += n : n === '"' || n === "'" ? t = n : n === ">" ? (c(), u = 1) : u && (n === "=" ? (u = 5, v = f, f = "") : n === "/" && (u < 5 || l[s][m + 1] === ">") ? (c(), u === 3 && (g = g[0]), u = g, (g = g[0]).push(2, 0, u), u = 0) : n === " " || n === "	" || n === `
-` || n === "\r" ? (c(), u = 2) : f += n), u === 3 && f === "!--" && (u = 4, g = g[0]);
-        }
-        return c(), g;
-    }(p)), o), arguments, [])).length > 1 ? o : o[0];
-}
-const html1 = w.bind(ye);
-export { html1 as html, xe as render, Ce as Component, U1 as useEffect, O1 as useLayoutEffect, Y1 as useState, z1 as useRef, G1 as useMemo, W1 as useCallback, me as createContext, ke as createRef, X1 as useContext, ye as h, ue as cloneElement, he as hydrate,  };
